Reworked multi world permission support so that the default permissions are loaded from config.yml.
This prevents users getting very confused when the rename their world and
appear to loose all of their permissions. It also makes it much more supportable
in a GSP environment.

Note: When starting on an existing install it ensures the old default permissions
<level-name>.yml is copied to config.yml

Cleaned up quite a bit of logic in com.nijiko.permissions.Control
* Refactored methods removing duplicate code
* Removed unrequired condition checks
* Removed unreachable statements